+++ CPP_Final_Project/CPP_Final_Project/SensorData.h
@@ -8,9 +8,9 @@ class SensorData
 public:
 	string _sensorName;
 	int _testNumber;
-	float _temperature;
+	int _temperature;
 
-	SensorData(string sensorName, int testNumber, float temperature);
+	SensorData(string sensorName, int testNumber, int temperature);
 
 	std::string ToString() const;
 };
